requirements.txt
requests~=2.32.3
PyYAML~=6.0.1
openai~=1.34.0
google~=3.0.0
SQLAlchemy~=2.0.31
pandas~=2.2.2
streamlit~=1.36.0
textblob~=0.18.0.post0
flask~=2.3.3
